move obstacle detection to sensors
