+++ obstacle_detection.py
@@ -28,26 +28,22 @@ class ObstacleDetector():
         
         self.obstacle = False
         self.obstacle_dir = 0
-        
-        self.depth_img = None
-        rospy.Subscriber('/camera/depth/image', Image, self._depthCallback)
 
-    def _getBlurredDepthSlice(self, min_height, max_height, min_width, max_width):
+    def _getBlurredDepthSlice(self, depth_img, min_height, max_height, min_width, max_width):
         """ Get a prepped slice of the image. """
-        return cv2.medianBlur(self.depth_img[min_height:max_height, min_width:max_width], 5)
+        return cv2.medianBlur(depth_img[min_height:max_height, min_width:max_width], 5)
     
     def _getIndex(self, img, operation):
         """ Get the index of the shortest distance in the depth image slice. """
         try:
             return np.unravel_index(operation(img[np.nonzero(img)]), img.shape)
         except ValueError:
-            self._logger.error("Encountered all NaN slice in depth image.")
             return None
     
-    def _extractObstacle(self):
+    def _extractObstacle(self, depth_img):
         """ If there is an obstacle nearby, find it. """
         # get slice to check distance on
-        img_height, img_width = self.depth_img.shape
+        img_height, img_width = depth_img.shape
         s_width = int(img_width * self._OBSTACLE_SAMPLE_WIDTH)
         w_center = img_width // 2
     
@@ -59,35 +55,12 @@ class ObstacleDetector():
         if min_index is None:
             # fail as safely as possible
             self.obstacle = True
-            return
+            return False
         
         # if the closest thing in our slice is too close, likely an obstacle
         if sample[min_index] < self._OBSTACLE_DIST_THRESH:
             if not self.obstacle:
                 self.obstacle_dir = -1 if min_index[1] < w_center else 1
                 self.obstacle = True
-                self._logger.warn("Encountered obstacle on the " + ["right.", "left."][min_index[1] < w_center])
         else:
             self.obstacle = False
-
-    def _depthCallback(self, data):
-        """ Process incoming depth data. """
-        # get the depth image
-        self.depth_img = self.bridge.imgmsg_to_cv2(data, 'passthrough')
-
-        # detect obstacles
-        self._extractObstacle()
-
-if __name__ == "__main__":
-    from tester import Tester
-
-    class ObstacleDetectorTest(Tester):
-        """ Behavioral tests for ObstacleDectection. """
-        def __init__(self):
-            self.obstacle_detector = ObstacleDetector()
-            Tester.__init__(self, "ObstacleDetection")
-
-        def main(self):
-            self.rate.sleep()
-
-    ObstacleDetectorTest()
